pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scientific-api"
version = "0.2.0"
description = "Scientific-API: cosmology ingestion + quantum graph ensembles + graph comparison (VKR)"
requires-python = ">=3.11,<3.12"
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "matplotlib",
  "pyarrow",
  "pyyaml",
  "httpx",
  "astropy",
  "tqdm",
  "networkx",
  "joblib",
]


[tool.setuptools.packages.find]
include = ["scientific_api*"]
